requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
diffusers==0.33.1
filelock==3.18.0
fsspec==2025.3.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.30.2
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.6.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pillow==11.2.1
python-telegram-bot==13.15
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tornado==6.1
tqdm==4.67.1
transformers==4.51.2
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
zipp==3.21.0
